requirements.txt
# PyTorch must be installed FIRST with the CUDA wheel — see README Setup section.
# Running `pip install -r requirements.txt` alone will give you CPU-only PyTorch.
# Run this before requirements.txt:
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchaudio>=2.0.0

openai-whisper>=20231117
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
tkinterdnd2>=0.3.0
Pillow>=10.0.0
customtkinter>=5.2.0

# LLM providers — install the ones you plan to use
anthropic>=0.40.0
openai>=1.0.0
google-genai>=1.0.0
